+++ src/mimic/src/orm/base/types/mod.rs
@@ -3,7 +3,6 @@ pub mod bytes;
 pub mod color;
 pub mod lang;
 pub mod math;
-pub mod orm;
 pub mod prim;
 pub mod text;
 pub mod time;