+++ src/mimic/src/orm/base/types/orm.rs
@@ -1,5 +1,0 @@
-///
-/// EntityPath
-///
-
-pub type EntityPath = String;