+++ src/mimic/src/orm/schema/node/field.rs
@@ -113,6 +113,11 @@ impl ValidateNode for Field {
             }
         }
 
+        // id
+        if self.name == "id" && self.value.default.is_none() {
+            errs.add("a field named id requires a default value");
+        }
+
         errs.result()
     }
 }