+++ src/mimic/src/orm/schema/types.rs
@@ -95,6 +95,7 @@ pub enum PrimitiveGroup {
     Float,
     Integer,
     String,
+    Ulid,
 }
 
 ///