+++ src/test/schema/src/admin.rs
@@ -8,7 +8,10 @@ use mimic::orm::{base::types, prelude::*};
     store = "crate::Store",
     sk(entity = "ComplexEntity", field = "id"),
     fields(
-        field(name = "id", value(item(is = "types::Ulid"))),
+        field(
+            name = "id",
+            value(item(is = "types::Ulid"), default = "types::Ulid::generate")
+        ),
         field(name = "string_test", value(item(is = "types::String"))),
         field(name = "principal_test", value(item(is = "types::Principal"))),
         field(name = "blob_test", value(item(is = "types::Blob"))),
@@ -67,7 +70,10 @@ pub struct ComplexEntity {}
     store = "crate::Store",
     sk(entity = "AdminEntity", field = "id"),
     fields(
-        field(name = "id", value(item(is = "types::Ulid"))),
+        field(
+            name = "id",
+            value(item(is = "types::Ulid"), default = "types::Ulid::generate")
+        ),
         field(name = "simple_text", value(item(is = "types::String"))),
         field(name = "tuple_test", value(item(is = "Tuple"))),
         field(name = "text_many", value(many, item(is = "types::String"))),
@@ -101,7 +107,10 @@ pub struct AdminEntity {}
     store = "crate::Store",
     sk(entity = "SimpleEntity", field = "id"),
     fields(
-        field(name = "id", value(item(is = "types::Ulid"))),
+        field(
+            name = "id",
+            value(item(is = "types::Ulid"), default = "types::Ulid::generate")
+        ),
         field(name = "name", value(item(is = "types::String"))),
     )
 )]
@@ -113,7 +122,10 @@ pub struct SimpleEntity {}
 
 #[record(
     fields(
-        field(name = "name", value(item(is = "types::String"))),
+        field(
+            name = "name",
+            value(item(is = "types::Ulid"), default = "types::Ulid::generate")
+        ),
         field(name = "variant_a", value(item(is = "EnumA"))),
         field(name = "description", value(item(is = "types::String"))),
     ),