+++ src/test/schema/src/store.rs
@@ -7,7 +7,10 @@ use mimic::orm::{base::types, prelude::*};
 #[entity(
     store = "crate::Store",
     sk(entity = "CreateBasic", field = "id"),
-    fields(field(name = "id", value(item(is = "types::Ulid"))))
+    fields(field(
+        name = "id",
+        value(item(is = "types::Ulid"), default = "types::Ulid::generate")
+    ))
 )]
 pub struct CreateBasic {}
 
@@ -19,7 +22,10 @@ pub struct CreateBasic {}
     store = "crate::Store",
     sk(entity = "Filterable", field = "id"),
     fields(
-        field(name = "id", value(item(is = "types::Ulid"))),
+        field(
+            name = "id",
+            value(item(is = "types::Ulid"), default = "types::Ulid::generate")
+        ),
         field(name = "name", value(item(is = "types::String"))),
         field(name = "description", value(item(is = "types::String"))),
     )
@@ -44,7 +50,10 @@ pub struct Limit {}
 #[entity(
     store = "crate::Store",
     sk(entity = "SortKeyOrder", field = "id"),
-    fields(field(name = "id", value(item(is = "types::Ulid"))))
+    fields(field(
+        name = "id",
+        value(item(is = "types::Ulid"), default = "types::Ulid::generate")
+    ))
 )]
 pub struct SortKeyOrder {}
 
@@ -55,7 +64,10 @@ pub struct SortKeyOrder {}
 #[entity(
     store = "crate::Store",
     sk(entity = "SortKeyA", field = "a_id"),
-    fields(field(name = "a_id", value(item(is = "types::Ulid"))))
+    fields(field(
+        name = "a_id",
+        value(item(is = "types::Ulid"), default = "types::Ulid::generate")
+    ))
 )]
 pub struct SortKeyA {}
 
@@ -86,7 +98,10 @@ pub struct SortKeyB {}
     fields(
         field(name = "a_id", value(item(is = "types::Ulid"))),
         field(name = "b_id", value(item(is = "types::Ulid"))),
-        field(name = "c_id", value(item(is = "types::Ulid"))),
+        field(
+            name = "c_id",
+            value(item(is = "types::Ulid"), default = "types::Ulid::generate")
+        )
     )
 )]
 pub struct SortKeyC {}