+++ src/test/schema/src/validate.rs
@@ -11,7 +11,10 @@ use mimic::orm::{
     store = "crate::Store",
     sk(entity = "ValidateTest", field = "id"),
     fields(
-        field(name = "id", value(item(is = "types::Ulid"))),
+        field(
+            name = "id",
+            value(item(is = "types::Ulid"), default = "types::Ulid::generate")
+        ),
         field(name = "multiple_ten", value(item(is = "MultipleTenType")))
     )
 )]